// sentio/new-morpho/src/schema/schema.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type { String, Int, BigInt, Float, ID, Bytes, Timestamp, Boolean, Int8 } from '@sentio/sdk/store'
import { Entity, Required, One, Many, Column, ListColumn, AbstractEntity, getStore, UpdateValues } from '@sentio/sdk/store'
import { BigDecimal } from '@sentio/bigdecimal'
import { DatabaseSchema } from '@sentio/sdk'







interface MorphoStateConstructorInput {
  id: ID;
  updatedAt: BigInt;
  vaultName: String;
  protocolId: String;
  collateralAsset: String;
  collateralDecimals: BigInt;
  collateralSymbol: String;
  collateralOracle: String;
  loanAsset: String;
  loanDecimals: BigInt;
  loanSymbol: String;
  collateralUsdPrice: BigDecimal;
  loanUsdPrice: BigDecimal;
  collateralUnderlyingApr: BigDecimal;
  loanUnderlyingApr: BigDecimal;
  borrowApy: BigDecimal;
  supplyApy: BigDecimal;
  loanToSharesRatio: BigDecimal;
  supplyToSharesRatio: BigDecimal;
  borrowRate: BigInt;
}
@Entity("MorphoState")
export class MorphoState extends AbstractEntity  {

	@Required
	@Column("ID")
	id: ID

	@Required
	@Column("BigInt")
	updatedAt: BigInt

	@Required
	@Column("String")
	vaultName: String

	@Required
	@Column("String")
	protocolId: String

	@Required
	@Column("String")
	collateralAsset: String

	@Required
	@Column("BigInt")
	collateralDecimals: BigInt

	@Required
	@Column("String")
	collateralSymbol: String

	@Required
	@Column("String")
	collateralOracle: String

	@Required
	@Column("String")
	loanAsset: String

	@Required
	@Column("BigInt")
	loanDecimals: BigInt

	@Required
	@Column("String")
	loanSymbol: String

	@Required
	@Column("BigDecimal")
	collateralUsdPrice: BigDecimal

	@Required
	@Column("BigDecimal")
	loanUsdPrice: BigDecimal

	@Required
	@Column("BigDecimal")
	collateralUnderlyingApr: BigDecimal

	@Required
	@Column("BigDecimal")
	loanUnderlyingApr: BigDecimal

	@Required
	@Column("BigDecimal")
	borrowApy: BigDecimal

	@Required
	@Column("BigDecimal")
	supplyApy: BigDecimal

	@Required
	@Column("BigDecimal")
	loanToSharesRatio: BigDecimal

	@Required
	@Column("BigDecimal")
	supplyToSharesRatio: BigDecimal

	@Required
	@Column("BigInt")
	borrowRate: BigInt
  constructor(data: MorphoStateConstructorInput) {super()}
  
  
  static update(values: UpdateValues<MorphoStateConstructorInput>): Promise<void> {
    return getStore().update(MorphoState, values)
  }
}


interface MetaVaultConstructorInput {
  id: ID;
  vault: String;
  vaultDecimals: BigInt;
  underlyingDecimals: BigInt;
  underlying: String;
  vaultName: String;
}
@Entity("MetaVault")
export class MetaVault extends AbstractEntity  {

	@Required
	@Column("ID")
	id: ID

	@Required
	@Column("String")
	vault: String

	@Required
	@Column("BigInt")
	vaultDecimals: BigInt

	@Required
	@Column("BigInt")
	underlyingDecimals: BigInt

	@Required
	@Column("String")
	underlying: String

	@Required
	@Column("String")
	vaultName: String
  constructor(data: MetaVaultConstructorInput) {super()}
  
  
  static update(values: UpdateValues<MetaVaultConstructorInput>): Promise<void> {
    return getStore().update(MetaVault, values)
  }
}


const source = `type MorphoState @entity {
  id: ID!
  updatedAt: BigInt!
  vaultName: String!
  protocolId: String!
  collateralAsset: String!
  collateralDecimals: BigInt!
  collateralSymbol: String!
  collateralOracle: String!
  loanAsset: String!
  loanDecimals: BigInt!
  loanSymbol: String!
  collateralUsdPrice: BigDecimal!
  loanUsdPrice: BigDecimal!
  collateralUnderlyingApr: BigDecimal!
  loanUnderlyingApr: BigDecimal!
  borrowApy: BigDecimal!
  supplyApy: BigDecimal!
  loanToSharesRatio: BigDecimal!
  supplyToSharesRatio: BigDecimal!
  borrowRate: BigInt!
}

type MetaVault @entity {
  id: ID!
  vault: String!
  vaultDecimals: BigInt!
  underlyingDecimals: BigInt!
  underlying: String!
  vaultName: String!
}
`
DatabaseSchema.register({
  source,
  entities: {
    "MorphoState": MorphoState,
		"MetaVault": MetaVault
  }
})